文件未生成的需要给出明确的提示

+++ application/tool/controller/FileExistsTraits.php
@@ -599,13 +599,20 @@ trait FileExistsTraits
         while(($html=readdir($resource))!=false){
             if(strpos($html,".html")!==false){
                 if($name.".html"==$html){
+                    if(!file_exists(ROOT_PATH."public/".$type."/".$html)){
+                        return json_encode([
+                            "msg"=>"文件未生成",
+                            "status"=>"failed",
+                            "data"=>''
+                        ]);
+                    }
                     $content=file_get_contents(ROOT_PATH."public/".$type."/".$html);
                 }
             }
         }
         return json_encode([
             "msg"=>"",
-            "stat"=>"success",
+            "status"=>"success",
             "data"=>$content
         ]);
     }
@@ -626,13 +633,20 @@ trait FileExistsTraits
         while(($html=readdir($resource))!=false){
             if(strpos($html,".html")!==false){
                 if($name.".html"==$html){
+                    if(!file_exists(ROOT_PATH."public/".$type."/".$html)){
+                        return json_encode([
+                            "msg"=>"文件未生成",
+                            "status"=>"failed",
+                            "data"=>''
+                        ]);
+                    }
                     $content=file_put_contents(ROOT_PATH."public/".$type."/".$html,$content);
                 }
             }
         }
         return json_encode([
             "msg"=>"",
-            "stat"=>"success",
+            "status"=>"success",
             "data"=>""
         ]);
     }
